Safari-611.1.10/LayoutTests/media/modern-media-controls/airplay-button/airplay-button-expected.txt

Testing the AirplayButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS airplayButton.element.localName is "button"
PASS airplayButton.element.classList.contains("airplay") is true
PASS airplayButton.iconName is Icons.Airplay
PASS airplayButton.image.element.style.webkitMaskImage.includes("macOS/Airplay.svg") became true
PASS macOSFullscreenAirplayButton.image.element.style.webkitMaskImage.includes("macOS/Airplay-fullscreen.svg") became true
PASS iOSAirplayButton.image.element.style.webkitMaskImage.includes("iOS/Airplay.svg") became true
PASS successfullyParsed is true

TEST COMPLETE
